+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/exception/HandlerNotFoundException.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.tobe.exception;
+package com.interface21.webmvc.servlet.mvc.exception;
 
 public class HandlerNotFoundException extends RuntimeException {
